requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.13
langchain-chroma>=0.0.1
langchain-text-splitters>=0.0.1
langchain-xai>=0.0.1
langgraph>=0.0.19
bs4>=0.0.1
beautifulsoup4>=4.12.2
chromadb>=0.4.22
typing-extensions>=4.5.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0 